fix: change password

+++ src/app/api/user/[id]/route.ts
@@ -1,10 +1,17 @@
 import { NextRequest, NextResponse } from 'next/server';
 import prisma from '../../../lib/db1';
 import { hash } from 'bcryptjs';
+import { checkUserRole } from 'utils/auth';
 
 //Get single user
 export async function GET(req: NextRequest, route: { params: { id: string } }) {
   try {
+    const allowedRoles = ['ADMIN'];
+    const hasrole = await checkUserRole(allowedRoles);
+    if (!hasrole) {
+      return NextResponse.json({ error: 'Access forbidden', status: 403 });
+    }
+
     const id = route.params.id;
     const user: Partial<User> = await prisma.user.findUnique({
       where: { id: id },
@@ -20,6 +27,12 @@ export async function GET(req: NextRequest, route: { params: { id: string } }) {
 //Update user
 export async function PUT(req: NextRequest, route: { params: { id: string } }) {
   try {
+    const allowedRoles = ['ADMIN'];
+    const hasrole = await checkUserRole(allowedRoles);
+    if (!hasrole) {
+      return NextResponse.json({ error: 'Access forbidden', status: 403 });
+    }
+
     const id = route.params.id;
     const result: User = await req.json();
     const { name, email, password, role, status } = result;
@@ -55,6 +68,12 @@ export async function DELETE(
   route: { params: { id: string } },
 ) {
   try {
+    const allowedRoles = ['ADMIN'];
+    const hasrole = await checkUserRole(allowedRoles);
+    if (!hasrole) {
+      return NextResponse.json({ error: 'Access forbidden', status: 403 });
+    }
+
     const id = route.params.id;
     const deletedUser = await prisma.user.delete({
       where: {
